# *: Improve the performance of the region tree