--- a/Gopkg.lock
+++ b/Gopkg.lock
@@ -197,8 +197,10 @@
 
 [[projects]]
   branch = "master"
+  digest = "1:05f95ffdfcf651bdb0f05b40b69e7f5663047f8da75c72d58728acb59b5cc107"
   name = "github.com/google/btree"
   packages = ["."]
+  pruneopts = "NUT"
   revision = "4030bb1f1f0c35b30ca7009e9ebd06849dd45306"
 
 [[projects]]
@@ -322,7 +324,7 @@
 
 [[projects]]
   branch = "master"
-  digest = "1:bb6d561c3d3ee23b72ce1feec757d6290fc9cc323e50675e0f0102b2669f17f9"
+  digest = "1:f25cdbc839cfc61f71b0f97fbb46fbdc741e951057060fc81cfc32591dd6e64e"
   name = "github.com/pingcap/kvproto"
   packages = [
     "pkg/eraftpb",
@@ -330,7 +332,7 @@
     "pkg/pdpb",
   ]
   pruneopts = "NUT"
-  revision = "5e6e69a5ed381bd4a8afe7cb96cc47f955f6d160"
+  revision = "e76f1c4d455b2f1b7373cc53d2176f5589dfc86e"
 
 [[projects]]
   digest = "1:8d8f554bbb62fb7aecf661b85b25e227f6ab6cfe2b4395ea65ef478bfc174940"