<!--
Thank you for working on PD! Please read PD's [CONTRIBUTING](https://github.com/pingcap/pd/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

## What have you changed? (required)
Our region tree is a B-tree, and we use the `iterate` to achieve the region-tree `update` and `find`. This PR mainly improve the B-tree `iterate` function, it will use the binary search to find the start item in the tree, reduces the time complexity from O(n) to O(log n) when positioning the start item. Therefore, reduced the consumption of cpu by comparing the region key.
<!--.
Please explain **IN DETAIL** what the changes are in this PR and why they are needed:
- Summarize your change (required)
- How does this PR work? Need a brief introduction for the changed logic (optional)
- Describe clearly one logical change and avoid lazy messages (optional)
- Describe any limitations of the current code (optional)

Please **NOTE** that:
- Do not assume the code is self-evident
- Do not assume reviewers understand the original issue
-->

## What are the type of the changes (required)?

<!--
The currently defined types are listed below, please pick one of the types for this PR by removing the others:
-->
- Improvement (non-breaking change which is an improvement to an existing feature)
)

## How has this PR been tested (required)?
<!--
Please describe the tests that you ran to verify your changes. Have you finished unit tests, integration tests, or manual tests?
-->
unit test
munaul test
## Benchmark result if necessary (optional)
before:
```
go test -bench="BenchmarkRegionTreeUpdate" -count 4 -run=none -benchtime=3s
goos: linux
goarch: amd64
pkg: github.com/pingcap/pd/server/core
BenchmarkRegionTreeUpdate-4 1000000 5361 ns/op
BenchmarkRegionTreeUpdate-4 1000000 5282 ns/op
BenchmarkRegionTreeUpdate-4 1000000 5387 ns/op
BenchmarkRegionTreeUpdate-4 1000000 5448 ns/op
 ```
```
(pprof) top10
Showing nodes accounting for 16190ms, 41.88% of 38660ms total
Dropped 448 nodes (cum <= 193.30ms)
Showing top 10 nodes out of 210
      flat  flat%   sum%        cum   cum%
    4860ms 12.57% 12.57%     4860ms 12.57%  runtime.cmpbody
    2120ms  5.48% 18.05%     2140ms  5.54%  runtime.heapBitsForObject
    1930ms  4.99% 23.05%     1930ms  4.99%  github.com/pingcap/pd/vendor/github.com/pingcap/kvproto/pkg/metapb.(*Region).GetStartKey (inline)
    1500ms  3.88% 26.93%     1500ms  3.88%  runtime.futex
    1300ms  3.36% 30.29%     5250ms 13.58%  runtime.scanobject
    1190ms  3.08% 33.37%     1280ms  3.31%  syscall.Syscall
     980ms  2.53% 35.90%     1100ms  2.85%  runtime.mapaccess2_fast64
     800ms  2.07% 37.97%      800ms  2.07%  runtime.usleep
     780ms  2.02% 39.99%     7610ms 19.68%  github.com/pingcap/pd/server/core.(*regionItem).Less
     730ms  1.89% 41.88%      730ms  1.89%  runtime.heapBits.bits (inline)
````
after:
```
go test -bench="BenchmarkRegionTreeUpdate" -count 4 -run=none -benchtime=3s
goos: linux
goarch: amd64
pkg: github.com/pingcap/pd/server/core
BenchmarkRegionTreeUpdate-4 2000000 2060 ns/op
BenchmarkRegionTreeUpdate-4 2000000 1921 ns/op
BenchmarkRegionTreeUpdate-4 2000000 1935 ns/op
BenchmarkRegionTreeUpdate-4 2000000 2002 ns/op
```
```
(pprof) top10
Showing nodes accounting for 13560ms, 33.43% of 40560ms total
Dropped 451 nodes (cum <= 202.80ms)
Showing top 10 nodes out of 232
      flat  flat%   sum%        cum   cum%
    2640ms  6.51%  6.51%     2670ms  6.58%  runtime.heapBitsForObject
    1910ms  4.71% 11.22%     6800ms 16.77%  runtime.scanobject
    1530ms  3.77% 14.99%     1700ms  4.19%  syscall.Syscall
    1450ms  3.57% 18.57%     1450ms  3.57%  runtime.futex
    1140ms  2.81% 21.38%     1140ms  2.81%  runtime.memmove
    1100ms  2.71% 24.09%     1100ms  2.71%  runtime.markBits.isMarked (inline)
    1050ms  2.59% 26.68%     1050ms  2.59%  runtime.cmpbody
    1020ms  2.51% 29.19%     1270ms  3.13%  runtime.mapaccess2_fast64
     950ms  2.34% 31.53%     3320ms  8.19%  runtime.mallocgc
     770ms  1.90% 33.43%      770ms  1.90%  runtime.usleep

```