@siddontang 2018-08-09
em, can we move the lock + b-tree to a lock free skiplist, see https://github.com/andy-kimball/arenaskl