@nolouch 2018-08-09
This lock-free SkipList base on the fixed-size space, not suitable for us. and now the upper bound on the number of comparisons, not on lock.
I bench it with single thread:
Btree value is `int`. Skiplist value is `int` to `bytes`:
Btree-Insert:
```
pkg: github.com/google/btree                        
BenchmarkInsertII-4      5000000               337 ns/op
BenchmarkInsertII-4      5000000               358 ns/op
BenchmarkInsertII-4      5000000               325 ns/op
BenchmarkInsertII-4      5000000               329 ns/op   
```
SkipList-Insert:
```
pkg: github.com/andy-kimball/arenaskl
BenchmarkInsert-4        3000000               410 ns/op
BenchmarkInsert-4        3000000               491 ns/op
BenchmarkInsert-4        2000000               573 ns/op
BenchmarkInsert-4        2000000               502 ns/op
```
Btree-Seek(Before improvement)：
```
pkg: github.com/google/btree
BenchmarkSeek-4          2000000               866 ns/op
BenchmarkSeek-4          2000000               865 ns/op
BenchmarkSeek-4          2000000               882 ns/op
BenchmarkSeek-4          2000000               917 ns/op
```
BTree-Seek(After improvement)：
```
pkg: github.com/google/btree
BenchmarkSeek-4          5000000               305 ns/op
BenchmarkSeek-4          5000000               297 ns/op
BenchmarkSeek-4          3000000               335 ns/op
BenchmarkSeek-4          5000000               307 ns/op
```
SkipList-Seek：
```
pkg: github.com/andy-kimball/arenaskl
BenchmarkSeek-4          5000000               285 ns/op
BenchmarkSeek-4          5000000               297 ns/op
BenchmarkSeek-4          5000000               314 ns/op
BenchmarkSeek-4          5000000               299 ns/op
```