@disksing 2018-08-13
I think you can try to merge the optimization into upstream `google/btree` repo.