@siddontang 2018-08-13
LGTM

Do you add more tests for btree now?