@nolouch 2018-08-13
@disksing  yeah, I already send a PR. @siddontang the tests already covered, I just add benchmark now.