@nolouch 2018-08-14
/run-all-tests